move helper functions into new file
